requirements.txt
# Core Dependencies
pyspark==3.5.0
delta-spark==3.0.0
kafka-python==2.0.2
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Data Processing & Analytics
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.2
fastparquet==2023.10.1

# Data Generation & Validation
faker==20.1.0
pydantic==2.5.2
jsonschema==4.20.0

# AWS Integration
boto3==1.34.0
botocore==1.34.0
s3fs==2023.12.2

# Database Connectors
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Configuration Management
pyyaml==6.0.1
python-dotenv==1.0.0
hydra-core==1.3.2

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
colorlog==6.8.0

# API & Dashboard Integration
requests==2.31.0
tableau-api-lib==0.1.48
msal==1.25.0  # Power BI authentication

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
testcontainers==3.7.1

# Development Tools
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.8.0

# Airflow (Optional - for workflow orchestration)
apache-airflow==2.8.1
apache-airflow-providers-amazon==8.15.0
apache-airflow-providers-snowflake==4.4.2

# Additional Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
schedule==1.2.0
